--- a/ADVISOR_ARCHITECTURE.md
+++ b/ADVISOR_ARCHITECTURE.md
@@ -1,0 +1,296 @@
+# 🎮 SMITE 2 Assault Advisor - Architecture & Deployment Options
+
+## 🏗️ Current Architecture
+
+### Desktop Application (Current Setup)
+The advisor is currently designed as a **desktop overlay application** that runs locally on your machine:
+
+```
+┌─────────────────────────────────────────────────────────────┐
+│                    SMITE 2 Game Window                     │
+│  ┌─────────────────────────────────────────────────────┐    │
+│  │              Assault Brain Overlay                  │    │
+│  │  🎯 WIN PROBABILITY: 78%                           │    │
+│  │  📊 TEAM ANALYSIS                                  │    │
+│  │  🛡️ BUILD SUGGESTIONS                              │    │
+│  │  🎤 Voice Coach: "Focus their Zeus!"               │    │
+│  └─────────────────────────────────────────────────────┘    │
+│                                                             │
+│  [Your SMITE 2 Assault Match]                              │
+└─────────────────────────────────────────────────────────────┘
+```
+
+### Core Components
+1. **Screen Capture**: Monitors SMITE 2 window for loading screens
+2. **OCR Engine**: Extracts team compositions from screenshots
+3. **Analysis Engine**: Processes team matchups and generates recommendations
+4. **Overlay UI**: Displays real-time analysis over the game
+5. **Voice Coach**: Provides audio guidance during matches
+6. **Data Scraper**: Keeps item/god data updated from web sources
+
+## 🌐 Deployment Options Comparison
+
+### Option 1: Desktop Overlay (Current)
+**How it works:**
+- Runs as a local Windows/Mac/Linux application
+- Overlays directly on top of SMITE 2
+- Real-time screen capture and analysis
+- No internet required for core functionality
+
+**Pros:**
+- ✅ Zero latency - instant analysis
+- ✅ Works offline after initial data download
+- ✅ Direct game integration with overlay
+- ✅ Voice coaching during gameplay
+- ✅ Privacy - no data sent to servers
+- ✅ No monthly costs or server maintenance
+
+**Cons:**
+- ❌ Requires installation on each machine
+- ❌ Platform-specific builds needed
+- ❌ Updates require manual installation
+- ❌ Limited sharing capabilities
+
+**Best for:** Individual players who want instant, private analysis
+
+### Option 2: Web Application
+**How it would work:**
+- Upload screenshots to web interface
+- Analysis performed on cloud servers
+- Results displayed in browser
+- Mobile-friendly interface
+
+**Pros:**
+- ✅ No installation required
+- ✅ Cross-platform compatibility
+- ✅ Easy sharing and collaboration
+- ✅ Automatic updates
+- ✅ Mobile access
+
+**Cons:**
+- ❌ Requires manual screenshot uploads
+- ❌ No real-time overlay integration
+- ❌ Internet dependency
+- ❌ Server costs and maintenance
+- ❌ Privacy concerns with uploaded data
+
+**Best for:** Casual analysis, team planning, mobile users
+
+### Option 3: Hybrid Approach
+**How it would work:**
+- Desktop app for real-time overlay
+- Web dashboard for detailed analysis
+- Cloud sync for data and settings
+- Mobile companion app
+
+**Pros:**
+- ✅ Best of both worlds
+- ✅ Real-time + detailed analysis
+- ✅ Cross-device synchronization
+- ✅ Team collaboration features
+
+**Cons:**
+- ❌ Most complex to develop
+- ❌ Higher maintenance costs
+- ❌ Multiple codebases to maintain
+
+## 🎯 Recommended Architecture: Enhanced Desktop + Web Dashboard
+
+### Primary: Enhanced Desktop Application
+
+```
+┌─────────────────────────────────────────────────────────────┐
+│                 SMITE 2 Assault Brain                      │
+│                    Desktop Application                     │
+├─────────────────────────────────────────────────────────────┤
+│                                                             │
+│  🎮 Real-Time Features:                                     │
+│  ├── Screen capture & OCR                                   │
+│  ├── Live overlay with analysis                             │
+│  ├── Voice coaching during matches                          │
+│  ├── Instant build suggestions                              │
+│  └── Win probability tracking                               │
+│                                                             │
+│  📊 Data Management:                                        │
+│  ├── Local SQLite database                                  │
+│  ├── Automatic web scraping                                 │
+│  ├── Cloud backup (optional)                               │
+│  └── Offline mode support                                   │
+│                                                             │
+│  🔗 Integrations:                                           │
+│  ├── Discord Rich Presence                                  │
+│  ├── Discord bot commands                                   │
+│  ├── Webhook notifications                                  │
+│  └── API for external tools                                 │
+│                                                             │
+└─────────────────────────────────────────────────────────────┘
+```
+
+### Secondary: Web Dashboard (Optional)
+
+```
+┌─────────────────────────────────────────────────────────────┐
+│              Web Dashboard (Optional)                      │
+│            https://smite-assault-brain.com                 │
+├─────────────────────────────────────────────────────────────┤
+│                                                             │
+│  📈 Analytics & History:                                    │
+│  ├── Match history and trends                               │
+│  ├── Performance analytics                                  │
+│  ├── God mastery tracking                                   │
+│  └── Improvement recommendations                            │
+│                                                             │
+│  👥 Team Features:                                          │
+│  ├── Team composition planning                              │
+│  ├── Shared analysis and builds                             │
+│  ├── Tournament preparation                                 │
+│  └── Coaching tools                                         │
+│                                                             │
+│  🔧 Management:                                             │
+│  ├── Settings synchronization                               │
+│  ├── Custom build templates                                 │
+│  ├── Data export/import                                     │
+│  └── Community features                                     │
+│                                                             │
+└─────────────────────────────────────────────────────────────┘
+```
+
+## 🚀 Implementation Roadmap
+
+### Phase 1: Enhanced Desktop App (Current Focus)
+1. **Improve Data Scraping**
+   - Real SmiteSource.com integration
+   - Tracker.gg API integration
+   - Automatic patch detection
+   - Local data caching optimization
+
+2. **Enhanced Analysis Engine**
+   - Machine learning for better predictions
+   - Situational build optimization
+   - Counter-pick suggestions
+   - Meta adaptation
+
+3. **Voice Coaching System**
+   - Text-to-speech integration
+   - Context-aware advice
+   - Multiple coaching personalities
+   - Customizable triggers
+
+4. **Discord Integration**
+   - Rich Presence updates
+   - Bot commands for analysis
+   - Webhook notifications
+   - Server integration features
+
+### Phase 2: Web Dashboard (Future)
+1. **Analytics Platform**
+   - Match history tracking
+   - Performance trends
+   - Improvement suggestions
+   - Comparative analysis
+
+2. **Team Collaboration**
+   - Shared team compositions
+   - Draft planning tools
+   - Tournament preparation
+   - Coaching features
+
+3. **Community Features**
+   - Build sharing
+   - Strategy discussions
+   - Meta analysis
+   - Educational content
+
+### Phase 3: Mobile Companion (Future)
+1. **Mobile App**
+   - Quick analysis tools
+   - Build references
+   - Team planning
+   - Notifications
+
+## 🔧 Technical Implementation Details
+
+### Desktop Application Stack
+```
+Frontend: Tkinter (Python) with custom theming
+Backend: Python with asyncio for performance
+Data: SQLite + compressed JSON for caching
+Vision: OpenCV + Tesseract/EasyOCR for screen analysis
+Voice: pyttsx3 for text-to-speech
+Networking: aiohttp for web scraping
+Packaging: PyInstaller for distribution
+```
+
+### Web Dashboard Stack (Future)
+```
+Frontend: React/Vue.js with responsive design
+Backend: FastAPI (Python) or Node.js
+Database: PostgreSQL with Redis caching
+Authentication: OAuth2 with Discord integration
+Hosting: Cloud platform (AWS/GCP/Azure)
+CDN: CloudFlare for global performance
+```
+
+### Data Flow Architecture
+```
+┌─────────────────┐    ┌─────────────────┐    ┌─────────────────┐
+│   SmiteSource   │    │   Tracker.gg    │    │  Official API   │
+│   Web Scraper   │    │   Integration   │    │   (Future)      │
+└─────────┬───────┘    └─────────┬───────┘    └─────────┬───────┘
+          │                      │                      │
+          └──────────────────────┼──────────────────────┘
+                                 │
+                    ┌─────────────▼─────────────┐
+                    │     Data Processing       │
+                    │   - Validation            │
+                    │   - Normalization         │
+                    │   - Analysis              │
+                    └─────────────┬─────────────┘
+                                 │
+                    ┌─────────────▼─────────────┐
+                    │    Local Database         │
+                    │   - SQLite storage        │
+                    │   - Compressed cache      │
+                    │   - Backup system         │
+                    └─────────────┬─────────────┘
+                                 │
+          ┌──────────────────────┼──────────────────────┐
+          │                      │                      │
+┌─────────▼───────┐    ┌─────────▼───────┐    ┌─────────▼───────┐
+│  Desktop App    │    │  Discord Bot    │    │  Web Dashboard  │
+│  - Real-time    │    │  - Commands     │    │  - Analytics    │
+│  - Overlay      │    │  - Rich Presence│    │  - Sharing      │
+│  - Voice Coach  │    │  - Webhooks     │    │  - Planning     │
+└─────────────────┘    └─────────────────┘    └─────────────────┘
+```
+
+## 🎯 Why Desktop-First Approach is Optimal
+
+### For SMITE 2 Assault Specifically:
+1. **Real-Time Requirements**: Assault matches are fast-paced, requiring instant analysis
+2. **Loading Screen Window**: Only ~30 seconds to analyze and provide recommendations
+3. **In-Game Coaching**: Voice guidance during team fights is crucial
+4. **Privacy**: Many players prefer not uploading game screenshots to web services
+5. **Reliability**: No internet dependency for core functionality
+
+### User Experience Benefits:
+- **Zero Friction**: No manual uploads or browser switching
+- **Instant Feedback**: Analysis appears immediately on loading screen
+- **Seamless Integration**: Feels like part of the game
+- **Always Available**: Works even with poor internet connection
+
+### Technical Advantages:
+- **Performance**: Local processing is faster than cloud round-trips
+- **Accuracy**: Direct screen capture is more reliable than manual uploads
+- **Customization**: Deep system integration for optimal user experience
+- **Cost Effective**: No server costs for core functionality
+
+## 🔮 Future Evolution Path
+
+1. **Start**: Desktop overlay with local analysis
+2. **Enhance**: Add Discord integration and voice coaching
+3. **Expand**: Optional web dashboard for analytics
+4. **Scale**: Mobile companion and team features
+5. **Integrate**: Official API partnerships and tournament tools
+
+This approach provides immediate value while building toward a comprehensive ecosystem that serves both casual and competitive players.