… ADVISOR_SETUP_GUIDE.md
… FINAL_SYSTEM_SUMMARY.md
… FUNCTIONALITY_OUTLINE.md
… OPTIMIZATION_ANALYSIS.md
… README_OPTIMIZED.md
… assault_brain_unified.py
… assault_data/unified.db
… data_backups/gods_20250528_035718.json.gz
… data_backups/items_20250528_035718.json.gz
… data_cache/smite_data.db
… demo_cool_features.py
… live_data/live_data.db
… optimized_assault_brain.py
… practical_data/assault_data.db
… requirements.txt
… smart_data/smart.db
… src/data/practical_data_manager.py
… src/data/real_web_scraper.py
… src/data/smite_data_scraper.py